refactor: collapse io and serializer into single class
