+++ README.md
@@ -58,8 +58,7 @@ import pytest
 @pytest.fixture
 def snapshot_custom(snapshot):
     return snapshot.with_class(
-        io_class=CustomIOClass,
-        serializer_class=CustomSerializerClass,
+        serializer_class=CustomIOClass,
     )
 
 def test_image(snapshot_custom):
@@ -67,7 +66,7 @@ def test_image(snapshot_custom):
     assert actual == snapshot_custom
 ```
 
-Both `CustomIOClass` and `CustomSerializerClass` should extend `syrupy.io.SnapshotIO` and `syrupy.io.SnapshotSerializer` respectively.
+`CustomIOClass` should extend `syrupy.serializer.SnapshotSerializer`.
 
 ## Uninstalling
